+++ src/components/index.js
@@ -9,3 +9,4 @@ export { default as StatusIconsLayout } from './status-icons-layout';
 export { default as ProgressBar } from './progress-bar';
 export { default as FilterBar } from './filter-bar';
 export { default as DatePicker } from './date-picker';
+export { default as TimePicker } from './time-picker';